feat: tighten up `bin` subdir handling
We currently strip a trailing `/bin` directory in several places. This was
done to work around a bug that shipped in cargo-dist between 0.10.0 and
0.12.0, where the prefix was always set to the `bin` directory instead of
the actual install path. This was introduced in
https://github.com/axodotdev/cargo-dist/commit/91ac6af587cf221b62d56c98cf99fb169c449011,
which first shipped in 0.10.0, and finally eliminated in the install path rewrite
that shipped in 0.12.0.

Always doing this actually introduced a few bugs of its own, so this now works around
it by actually checking the version of the cargo-dist that produced the install receipt
and only applying the correction if that version is within the 0.10.0 to pre-0.12.0
range. Luckily, we always included cargo-dist install info in the receipt from the
very first commit that introduced receipts, even though we weren't reading them until
now, so we're always able to perform this check.

Fixes #108.