fix: try to situate Windows exe tempdir in install_prefix
It's possible for the Windows tempdir to be on a different drive than
the currently running EXE, which makes `rename()` impossible. In case
that happens to us, we should handle it by creating the temp path to
rename the Windows exe to inside the install prefix, a path that we
should be able to trust to be a) writable to us, and b) on the same
drive as the EXE itself.

If we don't know the install_prefix, fall back to the old behaviour of
creating a temporary directory inside the system temp dir.

Fixes #120.