feat: return stdout/stderr from installers if able
We print the stdout/stderr to the terminal by default, but if the
caller has requested us to turn that off, we were instead throwing
that output away. This now instead copies that output to the error
type in the case that we didn't print it to the terminal, allowing
the caller to try to do something with it.